Starting refactor of FnApiRunner
