+++ sdks/python/apache_beam/runners/portability/fn_api_runner/fn_runner.py
@@ -35,7 +35,6 @@ import threading
 import time
 from builtins import object
 from typing import TYPE_CHECKING
-from typing import Any
 from typing import Callable
 from typing import DefaultDict
 from typing import Dict
@@ -54,10 +53,8 @@ from typing import cast
 from typing import overload
 
 import grpc
-from typing_extensions import Protocol
 
 import apache_beam as beam  # pylint: disable=ungrouped-imports
-from apache_beam import coders
 from apache_beam.coders.coder_impl import create_InputStream
 from apache_beam.coders.coder_impl import create_OutputStream
 from apache_beam.metrics import metric
@@ -79,7 +76,12 @@ from apache_beam.runners import pipeline_context
 from apache_beam.runners import runner
 from apache_beam.runners.portability import artifact_service
 from apache_beam.runners.portability import portable_metrics
+from apache_beam.runners.portability.fn_api_runner import execution
 from apache_beam.runners.portability.fn_api_runner import translations
+from apache_beam.runners.portability.fn_api_runner.execution import Buffer
+from apache_beam.runners.portability.fn_api_runner.execution import GroupingBuffer
+from apache_beam.runners.portability.fn_api_runner.execution import ListBuffer
+from apache_beam.runners.portability.fn_api_runner.execution import WindowGroupingBuffer
 from apache_beam.runners.portability.fn_api_runner.translations import create_buffer_id
 from apache_beam.runners.portability.fn_api_runner.translations import only_element
 from apache_beam.runners.portability.fn_api_runner.translations import split_buffer_id
@@ -91,20 +93,15 @@ from apache_beam.runners.worker.channel_factory import GRPCChannelFactory
 from apache_beam.runners.worker.sdk_worker import _Future
 from apache_beam.runners.worker.statecache import StateCache
 from apache_beam.transforms import environments
-from apache_beam.transforms import trigger
 from apache_beam.transforms.window import GlobalWindow
-from apache_beam.transforms.window import GlobalWindows
 from apache_beam.utils import profiler
 from apache_beam.utils import proto_utils
-from apache_beam.utils import windowed_value
 from apache_beam.utils.thread_pool_executor import UnboundedThreadPoolExecutor
 
 if TYPE_CHECKING:
   from apache_beam.pipeline import Pipeline
   from apache_beam.coders.coder_impl import CoderImpl
-  from apache_beam.coders.coder_impl import WindowedValueCoderImpl
   from apache_beam.portability.api import metrics_pb2
-  from apache_beam.transforms.window import BoundedWindow
 
 T = TypeVar('T')
 ConstructorFn = Callable[[
@@ -267,206 +264,6 @@ class BeamFnControlServicer(beam_fn_api_pb2_grpc.BeamFnControlServicer):
         'Runner: Requests multiplexing info: %s',
         [(str(req), worker)
          for req, worker in self._req_worker_mapping.items()])
-
-
-class Buffer(Protocol):
-  def __iter__(self):
-    # type: () -> Iterator[bytes]
-    pass
-
-  def append(self, item):
-    # type: (bytes) -> None
-    pass
-
-
-class PartitionableBuffer(Buffer, Protocol):
-  def partition(self, n):
-    # type: (int) -> List[List[bytes]]
-    pass
-
-
-class _ListBuffer():
-  """Used to support parititioning of a list."""
-  def __init__(self, coder_impl):
-    self._coder_impl = coder_impl
-    self._inputs = []  # type: List[bytes]
-    self._grouped_output = None
-    self.cleared = False
-
-  def append(self, element):
-    # type: (bytes) -> None
-    if self.cleared:
-      raise RuntimeError('Trying to append to a cleared ListBuffer.')
-    if self._grouped_output:
-      raise RuntimeError('ListBuffer append after read.')
-    self._inputs.append(element)
-
-  def partition(self, n):
-    # type: (int) -> List[List[bytes]]
-    if self.cleared:
-      raise RuntimeError('Trying to partition a cleared ListBuffer.')
-    if len(self._inputs) >= n or len(self._inputs) == 0:
-      return [self._inputs[k::n] for k in range(n)]
-    else:
-      if not self._grouped_output:
-        output_stream_list = [create_OutputStream() for _ in range(n)]
-        idx = 0
-        for input in self._inputs:
-          input_stream = create_InputStream(input)
-          while input_stream.size() > 0:
-            decoded_value = self._coder_impl.decode_from_stream(
-                input_stream, True)
-            self._coder_impl.encode_to_stream(
-                decoded_value, output_stream_list[idx], True)
-            idx = (idx + 1) % n
-        self._grouped_output = [[output_stream.get()]
-                                for output_stream in output_stream_list]
-      return self._grouped_output
-
-  def __iter__(self):
-    # type: () -> Iterator[bytes]
-    if self.cleared:
-      raise RuntimeError('Trying to iterate through a cleared ListBuffer.')
-    return iter(self._inputs)
-
-  def clear(self):
-    # type: () -> None
-    self.cleared = True
-    self._inputs = []
-    self._grouped_output = None
-
-
-class _GroupingBuffer(object):
-  """Used to accumulate groupded (shuffled) results."""
-  def __init__(self,
-               pre_grouped_coder,  # type: coders.Coder
-               post_grouped_coder,  # type: coders.Coder
-               windowing
-              ):
-    # type: (...) -> None
-    self._key_coder = pre_grouped_coder.key_coder()
-    self._pre_grouped_coder = pre_grouped_coder
-    self._post_grouped_coder = post_grouped_coder
-    self._table = collections.defaultdict(
-        list)  # type: DefaultDict[bytes, List[Any]]
-    self._windowing = windowing
-    self._grouped_output = None  # type: Optional[List[List[bytes]]]
-
-  def append(self, elements_data):
-    # type: (bytes) -> None
-    if self._grouped_output:
-      raise RuntimeError('Grouping table append after read.')
-    input_stream = create_InputStream(elements_data)
-    coder_impl = self._pre_grouped_coder.get_impl()
-    key_coder_impl = self._key_coder.get_impl()
-    # TODO(robertwb): We could optimize this even more by using a
-    # window-dropping coder for the data plane.
-    is_trivial_windowing = self._windowing.is_default()
-    while input_stream.size() > 0:
-      windowed_key_value = coder_impl.decode_from_stream(input_stream, True)
-      key, value = windowed_key_value.value
-      self._table[key_coder_impl.encode(key)].append(
-          value if is_trivial_windowing else windowed_key_value.
-          with_value(value))
-
-  def partition(self, n):
-    # type: (int) -> List[List[bytes]]
-
-    """ It is used to partition _GroupingBuffer to N parts. Once it is
-    partitioned, it would not be re-partitioned with diff N. Re-partition
-    is not supported now.
-    """
-    if not self._grouped_output:
-      if self._windowing.is_default():
-        globally_window = GlobalWindows.windowed_value(
-            None,
-            timestamp=GlobalWindow().max_timestamp(),
-            pane_info=windowed_value.PaneInfo(
-                is_first=True,
-                is_last=True,
-                timing=windowed_value.PaneInfoTiming.ON_TIME,
-                index=0,
-                nonspeculative_index=0)).with_value
-        windowed_key_values = lambda key, values: [
-            globally_window((key, values))]
-      else:
-        # TODO(pabloem, BEAM-7514): Trigger driver needs access to the clock
-        #   note that this only comes through if windowing is default - but what
-        #   about having multiple firings on the global window.
-        #   May need to revise.
-        trigger_driver = trigger.create_trigger_driver(self._windowing, True)
-        windowed_key_values = trigger_driver.process_entire_key
-      coder_impl = self._post_grouped_coder.get_impl()
-      key_coder_impl = self._key_coder.get_impl()
-      self._grouped_output = [[] for _ in range(n)]
-      output_stream_list = [create_OutputStream() for _ in range(n)]
-      for idx, (encoded_key, windowed_values) in enumerate(self._table.items()):
-        key = key_coder_impl.decode(encoded_key)
-        for wkvs in windowed_key_values(key, windowed_values):
-          coder_impl.encode_to_stream(wkvs, output_stream_list[idx % n], True)
-      for ix, output_stream in enumerate(output_stream_list):
-        self._grouped_output[ix] = [output_stream.get()]
-      self._table.clear()
-    return self._grouped_output
-
-  def __iter__(self):
-    # type: () -> Iterator[bytes]
-
-    """ Since partition() returns a list of lists, add this __iter__ to return
-    a list to simplify code when we need to iterate through ALL elements of
-    _GroupingBuffer.
-    """
-    return itertools.chain(*self.partition(1))
-
-
-class _WindowGroupingBuffer(object):
-  """Used to partition windowed side inputs."""
-  def __init__(
-      self,
-      access_pattern,
-      coder  # type: coders.WindowedValueCoder
-  ):
-    # type: (...) -> None
-    # Here's where we would use a different type of partitioning
-    # (e.g. also by key) for a different access pattern.
-    if access_pattern.urn == common_urns.side_inputs.ITERABLE.urn:
-      self._kv_extractor = lambda value: ('', value)
-      self._key_coder = coders.SingletonCoder('')  # type: coders.Coder
-      self._value_coder = coder.wrapped_value_coder
-    elif access_pattern.urn == common_urns.side_inputs.MULTIMAP.urn:
-      self._kv_extractor = lambda value: value
-      self._key_coder = coder.wrapped_value_coder.key_coder()
-      self._value_coder = (coder.wrapped_value_coder.value_coder())
-    else:
-      raise ValueError("Unknown access pattern: '%s'" % access_pattern.urn)
-    self._windowed_value_coder = coder
-    self._window_coder = coder.window_coder
-    self._values_by_window = collections.defaultdict(
-        list)  # type: DefaultDict[Tuple[str, BoundedWindow], List[Any]]
-
-  def append(self, elements_data):
-    # type: (bytes) -> None
-    input_stream = create_InputStream(elements_data)
-    while input_stream.size() > 0:
-      windowed_val_coder_impl = self._windowed_value_coder.get_impl(
-      )  # type: WindowedValueCoderImpl
-      windowed_value = windowed_val_coder_impl.decode_from_stream(
-          input_stream, True)
-      key, value = self._kv_extractor(windowed_value.value)
-      for window in windowed_value.windows:
-        self._values_by_window[key, window].append(value)
-
-  def encoded_items(self):
-    # type: () -> Iterator[Tuple[bytes, bytes, bytes]]
-    value_coder_impl = self._value_coder.get_impl()
-    key_coder_impl = self._key_coder.get_impl()
-    for (key, window), values in self._values_by_window.items():
-      encoded_window = self._window_coder.encode(window)
-      encoded_key = key_coder_impl.encode_nested(key)
-      output_stream = create_OutputStream()
-      for value in values:
-        value_coder_impl.encode_to_stream(value, output_stream, True)
-      yield encoded_key, encoded_window, output_stream.get()
 
 
 class FnApiRunner(runner.PipelineRunner):
@@ -702,16 +499,18 @@ class FnApiRunner(runner.PipelineRunner):
     metrics_by_stage = {}
     monitoring_infos_by_stage = {}
 
+    runner_execution_context = execution.FnApiRunnerExecutionContext(
+        worker_handler_manager.get_worker_handlers,
+        stage_context.components,
+        stage_context.safe_coders)
+
     try:
       with self.maybe_profile():
         pcoll_buffers = {}  # type: Dict[bytes, PartitionableBuffer]
         for stage in stages:
           stage_results = self._run_stage(
-              worker_handler_manager.get_worker_handlers,
-              stage_context.components,
-              stage,
-              pcoll_buffers,
-              stage_context.safe_coders)
+              runner_execution_context,
+              stage,)
           metrics_by_stage[stage.name] = stage_results.process_bundle.metrics
           monitoring_infos_by_stage[stage.name] = (
               stage_results.process_bundle.monitoring_infos)
@@ -721,23 +520,22 @@ class FnApiRunner(runner.PipelineRunner):
         runner.PipelineState.DONE, monitoring_infos_by_stage, metrics_by_stage)
 
   def _store_side_inputs_in_state(self,
-                                  worker_handler,  # type: WorkerHandler
-                                  context,  # type: pipeline_context.PipelineContext
-                                  pipeline_components,  # type: beam_runner_api_pb2.Components
+                                  bundle_context_manager,  # type: execution.BundleContextManager
+                                  runner_execution_context,  # type: execution.FnApiRunnerExecutionContext
                                   data_side_input,  # type: DataSideInput
-                                  pcoll_buffers,  # type: MutableMapping[bytes, PartitionableBuffer]
-                                  safe_coders
                                  ):
     # type: (...) -> None
     for (transform_id, tag), (buffer_id, si) in data_side_input.items():
       _, pcoll_id = split_buffer_id(buffer_id)
-      value_coder = context.coders[safe_coders[
-          pipeline_components.pcollections[pcoll_id].coder_id]]
-      elements_by_window = _WindowGroupingBuffer(si, value_coder)
-      if buffer_id not in pcoll_buffers:
-        pcoll_buffers[buffer_id] = _ListBuffer(
+      value_coder = bundle_context_manager.pipeline_context.coders[
+        runner_execution_context.safe_coders[
+          runner_execution_context.pipeline_components.pcollections[
+            pcoll_id].coder_id]]
+      elements_by_window = WindowGroupingBuffer(si, value_coder)
+      if buffer_id not in runner_execution_context.pcoll_buffers:
+        runner_execution_context.pcoll_buffers[buffer_id] = ListBuffer(
             coder_impl=value_coder.get_impl())
-      for element_data in pcoll_buffers[buffer_id]:
+      for element_data in runner_execution_context.pcoll_buffers[buffer_id]:
         elements_by_window.append(element_data)
 
       if si.urn == common_urns.side_inputs.ITERABLE.urn:
@@ -745,7 +543,8 @@ class FnApiRunner(runner.PipelineRunner):
           state_key = beam_fn_api_pb2.StateKey(
               iterable_side_input=beam_fn_api_pb2.StateKey.IterableSideInput(
                   transform_id=transform_id, side_input_id=tag, window=window))
-          worker_handler.state.append_raw(state_key, elements_data)
+          bundle_context_manager.worker_handler.state.append_raw(state_key,
+                                                                 elements_data)
       elif si.urn == common_urns.side_inputs.MULTIMAP.urn:
         for key, window, elements_data in elements_by_window.encoded_items():
           state_key = beam_fn_api_pb2.StateKey(
@@ -761,10 +560,9 @@ class FnApiRunner(runner.PipelineRunner):
   def _run_bundle_multiple_times_for_testing(
       self,
       worker_handler_list,  # type: Sequence[WorkerHandler]
-      process_bundle_descriptor,
+      bundle_context_manager,  # type: execution.BundleContextManager
       data_input,
       data_output,  # type: DataOutput
-      get_input_coder_callable,
       cache_token_generator
   ):
     # type: (...) -> None
@@ -780,9 +578,9 @@ class FnApiRunner(runner.PipelineRunner):
         testing_bundle_manager = ParallelBundleManager(
             worker_handler_list,
             lambda pcoll_id,
-            transform_id: _ListBuffer(coder_impl=get_input_coder_callable),
-            get_input_coder_callable,
-            process_bundle_descriptor,
+            transform_id: ListBuffer(coder_impl=bundle_context_manager.get_input_coder),
+            bundle_context_manager.get_input_coder,
+            bundle_context_manager.process_bundle_descriptor,
             self._progress_frequency,
             k,
             num_workers=self._num_workers,
@@ -793,10 +591,9 @@ class FnApiRunner(runner.PipelineRunner):
 
   def _collect_written_timers_and_add_to_deferred_inputs(
       self,
-      context,  # type: pipeline_context.PipelineContext
       pipeline_components,  # type: beam_runner_api_pb2.Components
       stage,  # type: translations.Stage
-      get_buffer_callable,
+      bundle_context_manager,  # type: execution.BundleContextManager
       deferred_inputs  # type: MutableMapping[str, PartitionableBuffer]
   ):
     # type: (...) -> None
@@ -804,9 +601,9 @@ class FnApiRunner(runner.PipelineRunner):
     for transform_id, timer_writes in stage.timer_pcollections:
 
       # Queue any set timers as new inputs.
-      windowed_timer_coder_impl = context.coders[
+      windowed_timer_coder_impl = bundle_context_manager.pipeline_context.coders[
           pipeline_components.pcollections[timer_writes].coder_id].get_impl()
-      written_timers = get_buffer_callable(
+      written_timers = bundle_context_manager.get_buffer(
           create_buffer_id(timer_writes, kind='timers'), transform_id)
       if not written_timers.cleared:
         # Keep only the "last" timer set per key and window.
@@ -825,7 +622,7 @@ class FnApiRunner(runner.PipelineRunner):
         for windowed_key_timer in timers_by_key_and_window.values():
           windowed_timer_coder_impl.encode_to_stream(
               windowed_key_timer, out, True)
-        deferred_inputs[transform_id] = _ListBuffer(
+        deferred_inputs[transform_id] = ListBuffer(
             coder_impl=windowed_timer_coder_impl)
         deferred_inputs[transform_id].append(out.get())
         written_timers.clear()
@@ -833,8 +630,7 @@ class FnApiRunner(runner.PipelineRunner):
   def _add_residuals_and_channel_splits_to_deferred_inputs(
       self,
       splits,  # type: List[beam_fn_api_pb2.ProcessBundleSplitResponse]
-      get_input_coder_impl_callable,
-      input_for_callable,
+      bundle_context_manager,  # type: execution.BundleContextManager
       last_sent,
       deferred_inputs  # type: MutableMapping[str, PartitionableBuffer]
   ):
@@ -843,15 +639,16 @@ class FnApiRunner(runner.PipelineRunner):
     prev_stops = {}  # type: Dict[str, int]
     for split in splits:
       for delayed_application in split.residual_roots:
-        name = input_for_callable(
+        name = bundle_context_manager.input_for(
             delayed_application.application.transform_id,
             delayed_application.application.input_id)
         if name not in deferred_inputs:
-          deferred_inputs[name] = _ListBuffer(
-              coder_impl=get_input_coder_impl_callable(name))
+          deferred_inputs[name] = ListBuffer(
+              coder_impl=bundle_context_manager.get_input_coder(name))
         deferred_inputs[name].append(delayed_application.application.element)
       for channel_split in split.channel_splits:
-        coder_impl = get_input_coder_impl_callable(channel_split.transform_id)
+        coder_impl = bundle_context_manager.get_input_coder(
+            channel_split.transform_id)
         # TODO(SDF): This requires determanistic ordering of buffer iteration.
         # TODO(SDF): The return split is in terms of indices.  Ideally,
         # a runner could map these back to actual positions to effectively
@@ -873,9 +670,9 @@ class FnApiRunner(runner.PipelineRunner):
             get(channel_split.transform_id, len(all_elements)) + 1]
         if residual_elements:
           if channel_split.transform_id not in deferred_inputs:
-            coder_impl = get_input_coder_impl_callable(
+            coder_impl = bundle_context_manager.get_input_coder(
                 channel_split.transform_id)
-            deferred_inputs[channel_split.transform_id] = _ListBuffer(
+            deferred_inputs[channel_split.transform_id] = ListBuffer(
                 coder_impl=coder_impl)
           deferred_inputs[channel_split.transform_id].append(
               coder_impl.encode_all(residual_elements))
@@ -910,11 +707,11 @@ class FnApiRunner(runner.PipelineRunner):
           else:
             coder = stage.context.coders[coder_id]
           if pcoll_id == translations.IMPULSE_BUFFER:
-            data_input[target] = _ListBuffer(coder_impl=coder.get_impl())
+            data_input[target] = ListBuffer(coder_impl=coder.get_impl())
             data_input[target].append(ENCODED_IMPULSE_VALUE)
           else:
             if pcoll_id not in pcoll_buffers:
-              data_input[target] = _ListBuffer(coder_impl=coder.get_impl())
+              data_input[target] = ListBuffer(coder_impl=coder.get_impl())
             data_input[target] = pcoll_buffers[pcoll_id]
         elif transform.spec.urn == bundle_processor.DATA_OUTPUT_URN:
           target = transform.unique_name, only_element(transform.inputs)
@@ -937,11 +734,8 @@ class FnApiRunner(runner.PipelineRunner):
     return data_input, data_side_input, data_output
 
   def _run_stage(self,
-                 worker_handler_factory,  # type: Callable[[Optional[str], int], List[WorkerHandler]]
-                 pipeline_components,  # type: beam_runner_api_pb2.Components
+                 runner_execution_context,  # type: execution.FnApiRunnerExecutionContext
                  stage,  # type: translations.Stage
-                 pcoll_buffers,  # type: MutableMapping[bytes, PartitionableBuffer]
-                 safe_coders
                 ):
     # type: (...) -> beam_fn_api_pb2.InstructionResponse
 
@@ -950,12 +744,7 @@ class FnApiRunner(runner.PipelineRunner):
     Args:
       worker_handler_factory: A ``callable`` that takes in an environment id
         and a number of workers, and returns a list of ``WorkerHandler``s.
-      pipeline_components (beam_runner_api_pb2.Components): TODO
       stage (translations.Stage)
-      pcoll_buffers (collections.defaultdict of str: list): Mapping of
-        PCollection IDs to list that functions as buffer for the
-        ``beam.PCollection``.
-      safe_coders (dict): TODO
     """
     def iterable_state_write(values, element_coder_impl):
       # type: (...) -> bytes
@@ -969,20 +758,20 @@ class FnApiRunner(runner.PipelineRunner):
           out.get())
       return token
 
-    worker_handler_list = worker_handler_factory(
+    worker_handler_list = runner_execution_context.worker_handler_factory(
         stage.environment, self._num_workers)
 
     # All worker_handlers share the same grpc server, so we can read grpc server
     # info from any worker_handler and read from the first worker_handler.
     worker_handler = next(iter(worker_handler_list))
     context = pipeline_context.PipelineContext(
-        pipeline_components, iterable_state_write=iterable_state_write)
+        runner_execution_context.pipeline_components, iterable_state_write=iterable_state_write)
     data_api_service_descriptor = worker_handler.data_api_service_descriptor()
 
     _LOGGER.info('Running %s', stage.name)
     data_input, data_side_input, data_output = self._extract_endpoints(
-        stage, pipeline_components, data_api_service_descriptor, pcoll_buffers,
-        context, safe_coders)
+        stage, runner_execution_context.pipeline_components, data_api_service_descriptor, runner_execution_context.pcoll_buffers,
+        context, runner_execution_context.safe_coders)
 
     process_bundle_descriptor = beam_fn_api_pb2.ProcessBundleDescriptor(
         id=self._next_uid(),
@@ -990,11 +779,14 @@ class FnApiRunner(runner.PipelineRunner):
             transform.unique_name: transform
             for transform in stage.transforms
         },
-        pcollections=dict(pipeline_components.pcollections.items()),
-        coders=dict(pipeline_components.coders.items()),
+        pcollections=dict(runner_execution_context.pipeline_components.pcollections.items()),
+        coders=dict(runner_execution_context.pipeline_components.coders.items()),
         windowing_strategies=dict(
-            pipeline_components.windowing_strategies.items()),
-        environments=dict(pipeline_components.environments.items()))
+            runner_execution_context.pipeline_components.windowing_strategies.items()),
+        environments=dict(runner_execution_context.pipeline_components.environments.items()))
+
+    bundle_context_manager = execution.BundleContextManager(
+        runner_execution_context, process_bundle_descriptor, worker_handler, context)
 
     state_api_service_descriptor = worker_handler.state_api_service_descriptor()
     if state_api_service_descriptor:
@@ -1004,92 +796,30 @@ class FnApiRunner(runner.PipelineRunner):
     # Store the required side inputs into state so it is accessible for the
     # worker when it runs this bundle.
     self._store_side_inputs_in_state(
-        worker_handler,
-        context,
-        pipeline_components,
-        data_side_input,
-        pcoll_buffers,
-        safe_coders)
-
-    def get_buffer(buffer_id, transform_id):
-      # type: (bytes, str) -> PartitionableBuffer
-
-      """Returns the buffer for a given (operation_type, PCollection ID).
-
-      For grouping-typed operations, we produce a ``_GroupingBuffer``. For
-      others, we produce a ``_ListBuffer``.
-      """
-      kind, name = split_buffer_id(buffer_id)
-      if kind in ('materialize', 'timers'):
-        if buffer_id not in pcoll_buffers:
-          pcoll_buffers[buffer_id] = _ListBuffer(
-              coder_impl=get_input_coder_impl(transform_id))
-        return pcoll_buffers[buffer_id]
-      elif kind == 'group':
-        # This is a grouping write, create a grouping buffer if needed.
-        if buffer_id not in pcoll_buffers:
-          original_gbk_transform = name
-          transform_proto = pipeline_components.transforms[
-              original_gbk_transform]
-          input_pcoll = only_element(list(transform_proto.inputs.values()))
-          output_pcoll = only_element(list(transform_proto.outputs.values()))
-          pre_gbk_coder = context.coders[safe_coders[
-              pipeline_components.pcollections[input_pcoll].coder_id]]
-          post_gbk_coder = context.coders[safe_coders[
-              pipeline_components.pcollections[output_pcoll].coder_id]]
-          windowing_strategy = context.windowing_strategies[
-              pipeline_components.pcollections[output_pcoll].
-              windowing_strategy_id]
-          pcoll_buffers[buffer_id] = _GroupingBuffer(
-              pre_gbk_coder, post_gbk_coder, windowing_strategy)
-      else:
-        # These should be the only two identifiers we produce for now,
-        # but special side input writes may go here.
-        raise NotImplementedError(buffer_id)
-      return pcoll_buffers[buffer_id]
-
-    def get_input_coder_impl(transform_id):
-      # type: (str) -> CoderImpl
-      coder_id = beam_fn_api_pb2.RemoteGrpcPort.FromString(
-          process_bundle_descriptor.transforms[transform_id].spec.payload
-      ).coder_id
-      assert coder_id
-      if coder_id in safe_coders:
-        return context.coders[safe_coders[coder_id]].get_impl()
-      else:
-        return context.coders[coder_id].get_impl()
+        bundle_context_manager,
+        runner_execution_context,
+        data_side_input)
 
     # Change cache token across bundle repeats
     cache_token_generator = FnApiRunner.get_cache_token_generator(static=False)
 
     self._run_bundle_multiple_times_for_testing(
         worker_handler_list,
-        process_bundle_descriptor,
+        bundle_context_manager,
         data_input,
         data_output,
-        get_input_coder_impl,
         cache_token_generator=cache_token_generator)
 
     bundle_manager = ParallelBundleManager(
         worker_handler_list,
-        get_buffer,
-        get_input_coder_impl,
+        bundle_context_manager.get_buffer,
+        bundle_context_manager.get_input_coder,
         process_bundle_descriptor,
         self._progress_frequency,
         num_workers=self._num_workers,
         cache_token_generator=cache_token_generator)
 
     result, splits = bundle_manager.process_bundle(data_input, data_output)
-
-    def input_for(transform_id, input_id):
-      # type: (str, str) -> str
-      input_pcoll = process_bundle_descriptor.transforms[transform_id].inputs[
-          input_id]
-      for read_id, proto in process_bundle_descriptor.transforms.items():
-        if (proto.spec.urn == bundle_processor.DATA_INPUT_URN and
-            input_pcoll in proto.outputs.values()):
-          return read_id
-      raise RuntimeError('No IO transform feeds %s' % transform_id)
 
     last_result = result
     last_sent = data_input
@@ -1102,26 +832,27 @@ class FnApiRunner(runner.PipelineRunner):
       deferred_inputs = {}  # type: Dict[str, PartitionableBuffer]
 
       self._collect_written_timers_and_add_to_deferred_inputs(
-          context, pipeline_components, stage, get_buffer, deferred_inputs)
+          runner_execution_context.pipeline_components, stage,
+          bundle_context_manager, deferred_inputs)
       # Queue any process-initiated delayed bundle applications.
       for delayed_application in last_result.process_bundle.residual_roots:
-        name = input_for(
+        name = bundle_context_manager.input_for(
             delayed_application.application.transform_id,
             delayed_application.application.input_id)
         if name not in deferred_inputs:
-          deferred_inputs[name] = _ListBuffer(
-              coder_impl=get_input_coder_impl(name))
+          deferred_inputs[name] = ListBuffer(
+              coder_impl=bundle_context_manager.get_input_coder(name))
         deferred_inputs[name].append(delayed_application.application.element)
       # Queue any runner-initiated delayed bundle applications.
       self._add_residuals_and_channel_splits_to_deferred_inputs(
-          splits, get_input_coder_impl, input_for, last_sent, deferred_inputs)
+          splits, bundle_context_manager, last_sent, deferred_inputs)
 
       if deferred_inputs:
         # The worker will be waiting on these inputs as well.
         for other_input in data_input:
           if other_input not in deferred_inputs:
-            deferred_inputs[other_input] = _ListBuffer(
-                coder_impl=get_input_coder_impl(other_input))
+            deferred_inputs[other_input] = ListBuffer(
+                coder_impl=bundle_context_manager.get_input_coder(other_input))
         # TODO(robertwb): merge results
         # TODO(BEAM-8486): this should be changed to _registered
         bundle_manager._skip_registration = True  # type: ignore[attr-defined]
@@ -1183,12 +914,12 @@ class FnApiRunner(runner.PipelineRunner):
           else:
             coder = context.coders[coder_id]
           if pcoll_id == translations.IMPULSE_BUFFER:
-            data_input[transform.unique_name] = _ListBuffer(
+            data_input[transform.unique_name] = ListBuffer(
                 coder_impl=coder.get_impl())
             data_input[transform.unique_name].append(ENCODED_IMPULSE_VALUE)
           else:
             if pcoll_id not in pcoll_buffers:
-              pcoll_buffers[pcoll_id] = _ListBuffer(coder_impl=coder.get_impl())
+              pcoll_buffers[pcoll_id] = ListBuffer(coder_impl=coder.get_impl())
             data_input[transform.unique_name] = pcoll_buffers[pcoll_id]
         elif transform.spec.urn == bundle_processor.DATA_OUTPUT_URN:
           data_output[transform.unique_name] = pcoll_id
